feat: move confirmation into the login
