+++ src/app/app.module.ts
@@ -24,7 +24,7 @@ import { AddReservationComponent } from './view/add-reservation/add-reservation.
 import { CertifiedDevicesDeveloperComponent } from './view/certified-devices-developer/certified-devices-developer.component';
 //import { TimezonePipe } from './utils/timezone.pipe';
 import { PipesModule } from './pipes.module';
-import { ConfirmemailComponent } from './view/confirmemail/confirmemail.component';
+import { ConfirmEmailComponent } from './view/confirmemail/confirmemail.component';
 import { ForgetPasswordComponent } from './view/forget-password/forget-password.component';
 import { ResetPasswordComponent } from './view/reset-password/reset-password.component';
 import { AllUsersComponent } from './view/all-users/all-users.component';
@@ -60,7 +60,7 @@ import { WithLoginUnverifiedLayoutComponent } from './nav/withloginunverifiedlay
     AddReservationComponent,
 
     CertifiedDevicesDeveloperComponent,
-    ConfirmemailComponent,
+    ConfirmEmailComponent,
     ForgetPasswordComponent,
     ResetPasswordComponent,
     AllUsersComponent,
